fix: 'GatherEdcElements' component
There was a problem with the former approach to
display education information. It wasn't possible
to store data.

Now, it's possible to store an object that represent
one educational institution into an array. So, each
block (set, collection) of information displayed
on screen correspond to a different school or university.